backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djoser==2.2.3
Pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.3
gunicorn==21.2.0
drf-spectacular==0.27.2
reportlab==4.2.5
